feat(initial commit): welcome to workflow-manager
Currently, workflow manager opens up an HTTP listening port and exposes the "/components" and "/id" route resources, the former for reporting a JSON summary of installed deis cluster components, the latter for returning the auto-generated UUIDv4 value that the running deis cluster will be identified by.

Over time, various cluster status, version freshness, and other proactive health assessment features will be added to this codebase. In its alpha state, workflow manager is more of a lightweight, k8s-aware HTTP abstraction, then it is a specific implementation. Managing deis workflow health is a major long-term initiative, look here for the fruits of that effort!

Jack and the Deis Team