fix(ci): CI pipeline now works for local+travis
Essentially, we were not properly tagging image builds with "deisci" during deploys. To remediate that, we've set "deisci" as the default image tag for all builds, which removes the need to re-run docker build commands during the deploy script.

Additionally, but not re-running docker build commands during deploy, we avoid the messiness of having a different execution CWD for `make`.

Restored the previous Dockerfile `CP . /` statement.